Move search bar component to common folder
